# adjust healthcheck timings