## Previous behavior:

- Healthcheck wasn't running for the first 30 seconds
- Every 10s a check was made, (with a 10s timeout of the check to finish)
- After 5 failed checks, container was marked unhealthy.

This means it was marked unhealthy after 30s + 50s = 1m20s.
This resulting in some slower systems, that the `docker compose up -d` to bail out when a container was marked unhealthy.

Yes, `detach` is **blocking** until all containers are **running**, and if a dependency is marked unhealthy, the last container wont launch.

---

## Example:

Nextcloud app has the following containers
- DB
- Redis
- Nextcloud
- Cron

Nextcloud depends on DB and Redis
Cron Depends on Nextcloud

DB and Redis is starting fairly quick.
Then Nextcloud starts which grabs a tar file and unpacks it. Which includes a HUGE amount of small files thus taking a long time on systems with HDDs. (btw a slow internet connection will increase the time as well, considering that the tar is multiple MBs)
If that takes more than 1m20s to download and unpack and start accepting requests. Nextcloud is marked unhealthy.

Cron is never started because Nextcloud is not healthy. So docker bails out and exits with non-zero code 

---

##  New behavior:

- Healthcheck wasn't running for the first 10 seconds
- Every 10s, will check health with `5s` timeout for it to complete
- After 30 failed attempts it will mark it unhealthy.

So that means the container has 10s + (10*30)s = 5m10s to become healthy.

The shorter `start_period` will give docker a faster feedback and hopefully make things a bit snappier.